drm/i915: component master at i915 driver load
A generic component master is added to hold the i915 registration
until all required kernel modules are up and active.

This is achieved through following steps:
	- moving the i915 driver registration to the component master's
		bind call
	- all required kernel modules will add one component each to
		component_match of I915 component master.

If no component is added to the I915 component master, due to CONFIG
selection or HW limitation, component master's bind call will be
triggered with no wait.

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
Suggested-by: Daniel Vetter <daniel.vetter@ffwll.ch>